refactor(project): move model graph extraction
