+++ crates/djls-project/src/lib.rs
@@ -1,5 +1,6 @@
 mod ast;
 mod db;
+mod models;
 mod names;
 mod parse;
 mod project;
@@ -59,7 +60,7 @@ pub use specs::ExtractedMessageTemplate;
 pub use templates::FilterArity;
 pub use templates::FilterArityMap;
 pub use specs::KnownOptions;
-pub use specs::ModelGraph;
+pub use models::ModelGraph;
 pub use specs::RequiredKeyword;
 pub use specs::SplitPosition;
 pub use templates::SymbolKey;
@@ -68,10 +69,10 @@ pub use specs::TagArgumentKind;
 pub use specs::TagRule;
 pub use specs::TagRuleMap;
 pub use templates::TemplateSymbolKind;
-pub use specs::compute_model_graph;
+pub use models::compute_model_graph;
 pub use specs::extract_block_specs;
 pub use specs::extract_filter_arities;
-pub use specs::extract_model_graph;
+pub use models::extract_model_graph;
 pub use specs::extract_tag_rules;
 pub use templates::InstalledSymbolCandidate;
 pub use templates::InstalledSymbolOrigin;
